Changes based on feedback from Ivan.

+++ docker/launch-app.sh
@@ -3,7 +3,8 @@
 keystoreLocation=${SERVER_SSL_KEYSTORE:?}
 keystorePassword=${SERVER_SSL_KEYSTOREPASSWORD:?}
 
-if [ -n "${TOKEN_CERT_PATH}" ] && [ -f "${TOKEN_CERT_PATH}" ]; then
+if [ -n "${TOKEN_CERT_PATH}" ] && [ -f "${TOKEN_CERT_PATH}" ] && [ -n "${TOKEN_KEY_PATH}" ] && [ -f "${TOKEN_KEY_PATH}" ]; then
+  echo "INFO: Found Token Cert and Key at ${TOKEN_CERT_PATH} and ${TOKEN_KEY_PATH}"
   # Build PEM file
   cat "${TOKEN_KEY_PATH}" > "$HOME/token.pem"
   cat "${TOKEN_CERT_PATH}" >> "$HOME/token.pem"
@@ -14,7 +15,8 @@ if [ -n "${TOKEN_CERT_PATH}" ] && [ -f "${TOKEN_CERT_PATH}" ]; then
 fi
 
 # Add cert for signing SAML communication with DOI
-if [ -n "${SAML_CERT_PATH}" ] && [ -f "${SAML_CERT_PATH}" ]; then
+if [ -n "${SAML_CERT_PATH}" ] && [ -f "${SAML_CERT_PATH}" ] && [ -n "${SAML_KEY_PATH}" ] && [ -f "${SAML_KEY_PATH}" ]; then
+  echo "INFO: Found SAML Signing Cert and Key at ${SAML_CERT_PATH} and ${SAML_KEY_PATH}"
   # Build PEM file
   cat "${SAML_KEY_PATH}" > "$HOME/saml.pem"
   cat "${SAML_CERT_PATH}" >> "$HOME/saml.pem"
@@ -24,10 +26,11 @@ if [ -n "${SAML_CERT_PATH}" ] && [ -f "${SAML_CERT_PATH}" ]; then
   keytool -v -importkeystore -deststorepass "$keystorePassword" -destkeystore "$keystoreLocation" -deststoretype PKCS12 -srckeystore "$HOME/saml.pkcs12" -srcstorepass "$keystorePassword" -srcstoretype PKCS12 -noprompt
 fi
 
-# Add cert served by the SAML IDP
-if [ -n "${samlIdpHost}" ]; then
-  openssl s_client -host $samlIdpHost -port $samlIdpPort -prexit -showcerts </dev/null | sed -ne '/-BEGIN CERTIFICATE-/,/-END CERTIFICATE-/p' > samlidp.crt;
-  keytool  -importcert -file samlidp.crt -alias samlidp -keystore "$keystoreLocation" -storepass "$keystorePassword" -noprompt;
+# Add cert served by the SAML IDP (never used if running in local dev mode)
+if [ -n "${samlIdpHost}" ] && [ $LOCAL_DEV_MODE = false ]; then
+  echo "INFO: Attempting to fetch and trust the SSL Certificate for the SAML IDP at: ${samlIdpHost}"
+  openssl s_client -host $samlIdpHost -port $samlIdpPort -prexit -showcerts </dev/null | sed -ne '/-BEGIN CERTIFICATE-/,/-END CERTIFICATE-/p' > /home/spring/samlidp.crt;
+  keytool  -importcert -file /home/spring/samlidp.crt -alias samlidp -keystore "$keystoreLocation" -storepass "$keystorePassword" -noprompt
 fi
 
 # Set correct Spring Launch Args if we are running local dev mode
@@ -42,4 +45,7 @@ fi
 # Most spring apps we build don't load the keystore via FileSystemResourceLoader, but Water Auth needs to for the SAML stuff
 PREFIXED_KEYSTORE_LOCATION="file:$keystoreLocation"
 
-java -server -Djava.security.egd=file:/dev/./urandom -DkeystoreLocation=$PREFIXED_KEYSTORE_LOCATION -DkeystorePassword=$keystorePassword -jar app.jar $SPRING_APP_ARGS "$@"
+# Export keystore password for java
+export keystorePassword=$keystorePassword
+
+java -server -Djava.security.egd=file:/dev/./urandom -DkeystoreLocation=$PREFIXED_KEYSTORE_LOCATION -jar app.jar $SPRING_APP_ARGS "$@"
